relocate prompt tester component
